дз 4

+++ controllers/contacts/removeContact.js
@@ -2,8 +2,11 @@ const { Contacts } = require('../../models/contacts');
 
 const removeContact = async(req, res, next)=> {
   try {
+    
     const { id } = req.params;
-    await Contacts.findOneAndDelete({_id: id});
+    const { _id: owner } = req.user;
+    await Contacts.findOneAndDelete({ _id: id, owner })
+    //await Contacts.findOneAndDelete({_id: id});
     res.status(200).json({ message: 'contact deleted' })
   } catch (error) {
     next()
